Add fix by ZeRo in order to remove warnings on FOFS() macros on file g_local.h.
Also add missing include of ctype.h on g_misc.c.

+++ src/g_local.h
@@ -41,6 +41,9 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 //
 #include "p_menu.h"
 
+// 2024 - intptr_t is more apropiated for compatibility with both 64 and 32 bits - ZeRo
+#include <stdint.h>
+
 // the "gameversion" client command will print this plus compile date
 #define	GAMEVERSION	"dday"
 //#define DEVVERSION	"4.1"
@@ -629,10 +632,11 @@ extern	int	meansOfDeath;
 
 extern	edict_t			*g_edicts;
 
-#define	FOFS(x) (int)&(((edict_t *)0)->x)
-#define	STOFS(x) (int)&(((spawn_temp_t *)0)->x)
-#define	LLOFS(x) (int)&(((level_locals_t *)0)->x)
-#define	CLOFS(x) (int)&(((gclient_t *)0)->x)
+// 2024 - Modified from int to intptr_t - ZeRo
+#define	FOFS(x) (intptr_t)&(((edict_t *)0)->x)
+#define	STOFS(x) (intptr_t)&(((spawn_temp_t *)0)->x)
+#define	LLOFS(x) (intptr_t)&(((level_locals_t *)0)->x)
+#define	CLOFS(x) (intptr_t)&(((gclient_t *)0)->x)
 
 #define random()	((rand () & 0x7fff) / ((float)0x7fff))
 #define crandom()	(2.0 * (random() - 0.5))
